Count perf optimizations.

+++ src/common.rs
@@ -4,15 +4,14 @@ use serde::{Deserialize, Serialize};
 macro_rules! registers_impls {
     ($len:expr, $ident:ident) => {
         // A Registers struct.
-        //
-        // Contains a `count` and a number of fixed size registers
-        // packed into `u32` integers.
         #[derive(Clone, Debug, Serialize, Deserialize)]
         pub struct $ident {
             // A buffer containing registers.
             buf:   Vec<u32>,
             // The number of registers stored in buf.
             count: usize,
+            // The number of registers set to zero.
+            zeros: usize,
         }
 
         impl $ident {
@@ -28,6 +27,7 @@ macro_rules! registers_impls {
                 $ident {
                     buf:   vec![0; ceil(count, Self::COUNT_PER_WORD)],
                     count: count,
+                    zeros: count,
                 }
             }
 
@@ -45,6 +45,7 @@ macro_rules! registers_impls {
             }
 
             #[inline] // Returns the value of the Register at `index`.
+            #[allow(dead_code)]
             pub fn get(&self, index: usize) -> u32 {
                 let (qu, rm) = (
                     index / Self::COUNT_PER_WORD,
@@ -54,17 +55,32 @@ macro_rules! registers_impls {
                 (self.buf[qu] >> (rm * Self::SIZE)) & Self::MASK
             }
 
-            #[inline] // Sets the value of the Register at `index` to `value`.
-            pub fn set(&mut self, index: usize, value: u32) {
+            #[inline] // Sets the value of the Register at `index` to `value`,
+                      // if `value` is greater than its current value.
+            pub fn set_greater(&mut self, index: usize, value: u32) {
                 let (qu, rm) = (
                     index / Self::COUNT_PER_WORD,
                     index % Self::COUNT_PER_WORD,
                 );
 
-                let mask = Self::MASK << (rm * Self::SIZE);
+                let cur = (self.buf[qu] >> (rm * Self::SIZE)) & Self::MASK;
 
-                self.buf[qu] =
-                    (self.buf[qu] & !mask) | (value << (rm * Self::SIZE));
+                if value > cur {
+                    if cur == 0 {
+                        self.zeros -= 1;
+                        self.buf[qu] |= (value << (rm * Self::SIZE));
+                    } else {
+                        let mask = Self::MASK << (rm * Self::SIZE);
+
+                        self.buf[qu] = (self.buf[qu] & !mask) |
+                            (value << (rm * Self::SIZE));
+                    }
+                }
+            }
+
+            #[inline]
+            pub fn zeros(&self) -> usize {
+                self.zeros
             }
 
             #[inline] // Returns the size of the Registers in bytes
@@ -83,6 +99,28 @@ registers_impls![5, Registers];
 // Registers implementation for 6-bit registers,
 // used by HyperLogLog++ implementation.
 registers_impls![6, RegistersPlus];
+
+// An array containing all possible values used to calculate
+// the "raw" sum.
+//
+// Instead of computing those values every time, look them up here.
+//
+// This is used only in the case the `const-loop` feature is enabled,
+// it requires a Rust compiler version 1.45.0 or higher.
+#[cfg(feature = "const-loop")]
+const RAW: [f64; 1 << RegistersPlus::SIZE] = {
+    const COUNT: usize = 1 << RegistersPlus::SIZE;
+
+    let mut raw = [0.0; COUNT];
+
+    let mut i = 0;
+    while i < COUNT {
+        raw[i] = 1.0 / (1u64 << i) as f64;
+        i += 1;
+    }
+
+    raw
+};
 
 // A trait for sharing common HyperLogLog related functionality between
 // different HyperLogLog implementations.
@@ -105,6 +143,30 @@ pub trait HyperLogLogCommon {
         raw = Self::alpha(count) * (count * count) as f64 / raw;
 
         (raw, zeros)
+    }
+
+    #[cfg(not(feature = "const-loop"))]
+    #[inline] // Returns the "raw" HyperLogLog estimate as defined by
+              // P. Flajolet et al. for a given `precision`.
+    fn estimate_raw_plus<I>(registers: I, count: usize) -> f64
+    where
+        I: Iterator<Item = u32>,
+    {
+        let raw: f64 = registers.map(|val| 1.0 / (1u64 << val) as f64).sum();
+
+        Self::alpha(count) * (count * count) as f64 / raw
+    }
+
+    #[cfg(feature = "const-loop")]
+    #[inline] // Returns the "raw" HyperLogLog estimate as defined by
+              // P. Flajolet et al. for a given `precision`.
+    fn estimate_raw_plus<I>(registers: I, count: usize) -> f64
+    where
+        I: Iterator<Item = u32>,
+    {
+        let raw: f64 = registers.map(|val| RAW[val as usize]).sum();
+
+        Self::alpha(count) * (count * count) as f64 / raw
     }
 
     #[inline] // Estimates the count of distinct elements using linear
@@ -164,13 +226,47 @@ mod tests {
 
         assert_eq!(registers.buf.len(), 2);
 
-        registers.set(1, 0b11);
+        registers.set_greater(1, 0b11);
 
         assert_eq!(registers.buf, vec![0b11000000, 0]);
 
-        registers.set(9, 0x7);
+        registers.set_greater(9, 0x7);
 
         assert_eq!(registers.buf, vec![0b11000000, 0x07000000]);
+    }
+
+    #[test]
+    fn test_registers_set_greater() {
+        let mut registers: RegistersPlus = RegistersPlus::with_count(10);
+
+        assert_eq!(registers.buf.len(), 2);
+
+        assert_eq!(registers.zeros(), 10);
+
+        registers.set_greater(1, 0);
+
+        assert_eq!(registers.buf, vec![0, 0]);
+        assert_eq!(registers.zeros(), 10);
+
+        registers.set_greater(1, 0b11);
+
+        assert_eq!(registers.buf, vec![0b11000000, 0]);
+        assert_eq!(registers.zeros(), 9);
+
+        registers.set_greater(9, 0x7);
+
+        assert_eq!(registers.buf, vec![0b11000000, 0x07000000]);
+        assert_eq!(registers.zeros(), 8);
+
+        registers.set_greater(1, 0b10);
+
+        assert_eq!(registers.buf, vec![0b11000000, 0x07000000]);
+        assert_eq!(registers.zeros(), 8);
+
+        registers.set_greater(9, 0x9);
+
+        assert_eq!(registers.buf, vec![0b11000000, 0x09000000]);
+        assert_eq!(registers.zeros(), 8);
     }
 
     #[test]
